+++ packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/next/objects/asset/Asset.kt
@@ -7,6 +7,7 @@ import expo.modules.medialibrary.next.objects.asset.delegates.AssetDelegate
 import expo.modules.medialibrary.next.objects.wrappers.RelativePath
 import expo.modules.medialibrary.next.objects.wrappers.MediaType
 import expo.modules.medialibrary.next.objects.wrappers.MimeType
+import expo.modules.medialibrary.next.records.AssetInfo
 import expo.modules.medialibrary.next.records.Location
 import expo.modules.medialibrary.next.records.Shape
 import kotlinx.coroutines.Dispatchers
@@ -41,6 +42,9 @@ class Asset(val assetDelegate: AssetDelegate) : SharedObject() {
 
   suspend fun getUri(): Uri =
     assetDelegate.getUri()
+
+  suspend fun getInfo(): AssetInfo =
+    assetDelegate.getInfo()
 
   suspend fun getMimeType(): MimeType =
     assetDelegate.getMimeType()